Call container.Terminate() on shutdown timeouts
We were logging if HcsShutdownComputeSystem failed, but we weren't
trying to force kill the container via Terminate after if we timed
out waiting for it to complete. Shutdown is async and we wait for
a notification for success, so most of the time the call itself will
return nil, but it doesn't indicate indicate success until we can
see that the system exited. So now we will fallback to Terminate for:

1. Shutdown returning an error that doesn't indicate the result is
to be waited on.
2. The async result of shutdown was non-nil
3. Waiting for the result passed the timeout we set.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>
(cherry picked from commit cae120b42536adea72101e6a358183c74a0bdac5)
Signed-off-by: Kevin Parsons <kevpar@microsoft.com>