gcs: Support setting sysctls from annotations when starting container
This change adds support for a new set of annotations:
  io.microsoft.container.sysctl.*
This causes gcs to set the sysctl at the end of the name with the value
of the annotation. For instance, if you set:
  io.microsoft.container.sysctl.net.core.rmem_default = 12345
Then the sysctl net.core.rmem_default will be set to the value 12345.

This may not be the correct long-term approach. We may not want to allow
any container with this annotation to set things (perhaps only
privileged containers, or the pause container in a pod). Or perhaps this
should be a separate bridge message to gcs, or part of initial bridge
configuration. However, this approach is quick and easy to implement,
which lets us add support here, and decide if we want to change it
later.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>