gcs: Add SCSIDevice type with remove operation
SCSI devices must be unplugged by the guest before removal on the host
side, to ensure smooth operation. Previously a SCSI device was unplugged
when a LCOWMappedVirtualDisk entry was removed. However, we want to
support multiple mounts per disk, which means we need to decouple unplug
from unmount.

This change introduces a new SCSIDevice resource type that has a remove
operation that can be used by the host to trigger an explicit unplug via
SCSI.

This is a breaking change to the bridge protocol:
- With new host/old guest, the host will attempt a SCSIDevice remove
  which will fail due to being unsupported by the guest.
- With old host/new guest, the host will expect the device to be
  unplugged when the disk is removed, which will no longer occur.